comment fixes
